% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{addIntercept}
\alias{addIntercept}
\title{Internal function to add a column for the intercept to a matrix.}
\usage{
addIntercept(x)
}
\arguments{
\item{x}{A matrix or data.frame.}
}
\value{
If `x` has no intercept, a matrix with an intercept column
}
\description{
Given a matrix or data.frame of size nxp, returns an nx(p+1)
object where the first column is an intercept; if the intercept is
not already in the object (defined as a first column being all
1's).
}
\author{
Josh Errickson
}

