% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbph.R
\name{confint.pbph}
\alias{confint.pbph}
\title{Confidence intervals for \code{pbph} object}
\usage{
\method{confint}{pbph}(object, parm, level = 0.95, ..., wald.style = FALSE,
  forceDisplayConfInt = FALSE, returnShape = FALSE)
}
\arguments{
\item{object}{An object of class \code{pbph}.}

\item{parm}{Parameters}

\item{level}{Confidence level.}

\item{...}{Additional arguments to \code{confint.lm}.}

\item{wald.style}{Logical. Defaults to \code{FALSE} so the method
performs test inversion to obtain proper coverage. If
\code{TRUE}, generates wald-style CI's, which will likely suffer
from undercoverage.}

\item{forceDisplayConfInt}{Logical. Defaults to \code{FALSE}. If a
hypothesis test for the interaction term fails, a confidence
interval is unreliable (and ultimately unnecessary) and thus not
returned. Set this to \code{TRUE} to force computation of the
interval regardless of the hypothesis test.}

\item{returnShape}{Logical. Defaults to \code{FALSE}. If \code{TRUE},
adds an attribute to the output the defines the shape of the
confidence region. If \code{TRUE}, \code{forceDisplayConfInt} is
set to \code{TRUE} as well.}
}
\value{
Confidence intervals
}
\description{
Similar to \code{confint.lm}.
}

