% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pblm-internal.R
\name{correctedvar}
\alias{correctedvar}
\title{(Internal) Compute B^-1*M*B^T}
\usage{
correctedvar(b11, b21, b22, m11, m22)
}
\arguments{
\item{b11, b21, b22, m11, m22}{Pieces of bread and meat.}
}
\value{
Variance estimate.
}
\description{
(Internal) Compute B^-1*M*B^T
}
\author{
Josh Errickson
}

