% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbph-internal.R
\name{corrVar}
\alias{corrVar}
\title{(Internal) Return covariance matrix associated with a given choice of eta.}
\usage{
corrVar(
  eta,
  object,
  breadAndMeat = createBreadAndMeat(object, cluster = object$pbph$cluster)
)
}
\arguments{
\item{eta}{Value of eta to use.}

\item{object}{A \code{pbph} object.}

\item{breadAndMeat}{By default, will create \code{breadAndMeat}
associated with \code{object}. If speed a concern, it is faster
to compute this once and pass into corrVar.}
}
\value{
A covariance \code{matrix}.
}
\description{
For a given eta, the covariance can vary. For example, etahat or
eta_0.
}
