% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pblm.R
\name{createBreadAndMeat}
\alias{createBreadAndMeat}
\title{Computes Bread and Meat matrices.}
\usage{
createBreadAndMeat(object)
}
\arguments{
\item{object}{A pblm object.}
}
\value{
A list of b11, b22, m11, and m22.
}
\description{
Computes the pieces of the Bread and Meat which do not depend on
eta. (e.g. all but B21)
}
\author{
Josh Errickson
}

