% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eottest}
\alias{eottest}
\title{Student level demographic, performance and end-of-term test scores.}
\format{A data.frame with 80 rows and 4 columns.
\itemize{
  \item afterschool Indicator in membership in an afterschool intervention (the treatment variable).
  \item male Indicator for gender (1 = male, 0 = female).
  \item gpa GPA on a 4.0 scale.
  \item test Score (out of 100) on an end-of-term standardized test.
}}
\usage{
eottest
}
\description{
This generated data set simulates some student-level data for
demonstrating the use of `pbph`.
}
\keyword{dataset}

