% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pblm.R
\name{hypothesisTest}
\alias{hypothesisTest}
\title{Conducts a hypothesis test for a given null.}
\usage{
hypothesisTest(object, null = 0)
}
\arguments{
\item{object}{A pblm object.}

\item{null}{Defaults to 0.}
}
\value{
A test statistic, with distribution t(k) where k is the
  number of parameters in the first stage model, less 2.
}
\description{
Conducts a hypothesis test for a given null.
}
\author{
Josh Errickson
}

