% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makemod2.R
\name{makemod2}
\alias{makemod2}
\title{(Internal) Fit the second stage of the Peters-Belson with
Prognostic Heterogeneity model.}
\usage{
makemod2(mod1, treatment, data, center = TRUE)
}
\arguments{
\item{mod1}{First stage model fit}

\item{treatment}{A vector of treatment statuses, should be all
\code{0} or \code{1}.}

\item{data}{Data where variables in \code{form} live.}

\item{center}{Default \code{TRUE}. Should the predicted values be
centered in the second stage?}
}
\value{
A \code{list} consisting of \code{mod2}, the second stage
  model, and \code{pred}, the predicted values from \code{mod1}.
}
\description{
Given a first stage model fit only on the control group, fit a second stage
model with formula.
}
\details{
resp - pred ~ pred | treatment == 1
}
