% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{makeSaveMatrix}
\alias{makeSaveMatrix}
\title{Generate an empty matrix suitable for saving simulation results.}
\usage{
makeSaveMatrix(colNames, reps)
}
\arguments{
\item{colNames}{Vector of column names.}

\item{reps}{Total reps.}
}
\value{
An empty matrix.
}
\description{
Given a vector of column names and a number of reps, generates an
appropriately sized matrix with given column names.
}
\author{
Josh Errickson
}

