% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{malldata}
\alias{malldata}
\title{Mall data and summer vs winter performance.}
\format{A data.frame with 238 rows and 4 columns.
\itemize{
  \item coveredgarage Indicator of the garage being covered as oppposed to an open lot (the treatment variable).
  \item year Year of construction.
  \item size Mall dimensions (in thousands square feets).
  \item winterhigher Binary indicator whether winter sales are higher than summer sales.
}}
\usage{
malldata
}
\description{
This generated data set simulates some data about malls to determine whether
having covered parking lots results in higher sales in winter over summer.
}
\keyword{dataset}

