% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pblm.R
\name{pblm}
\alias{pblm}
\title{Enhanced Peters-Belson method}
\usage{
pblm(mod1, treatment, data, center = TRUE, cluster = NULL)
}
\arguments{
\item{mod1}{First stage fitted model.}

\item{treatment}{Vector of 0/1 treatment indicators.}

\item{data}{Data where variables in \code{form} live.}

\item{center}{Default \code{TRUE}. Should the predicted values be
centered in the second stage?}

\item{cluster}{A \code{vector} defining the clusters.}
}
\value{
A \code{pblm} object which extends \code{lm}. Can be
  passed to \code{summary} or \code{confint}.

  The return contains an additional object, \code{epb}, which
  \code{lm} doesn't include. \code{epb} is a list which contains
  several pieces necessary for \code{summary} and \code{confint}
  support.
}
\description{
Performs enhanced Peters-Belson on the data.
}

