% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbph-package.R, R/pbph.R
\docType{package}
\name{pbph}
\alias{pbph}
\alias{pbph-package}
\title{pbph: A package for performing Peters-Belson with Prognostic Heterogeneity}
\usage{
pbph(mod1, treatment, data, center = TRUE, cluster = NULL)
}
\arguments{
\item{mod1}{First stage fitted model.}

\item{treatment}{A 0/1 variable in \code{data} indicating treatment
assignment, or a vector of 0/1 treatment indicators.}

\item{data}{Data where variables in \code{form} live.}

\item{center}{Default \code{TRUE}. Should the predicted values be
centered in the second stage?}

\item{cluster}{A variable in \code{data} indicating cluster
membership, or a vector of cluster membership.}
}
\value{
A \code{pbph} object which extends \code{lm}. Can be
  passed to \code{summary} or \code{confint}.

  The return contains an additional object, \code{pbph}, which \code{lm}
  doesn't include. \code{pbph} contains \code{mod1}, a copy of the first
  stage model, and \code{data}, which includes the \code{data} augmented
  with \code{treatment} and \code{predicted}.
}
\description{
pbph: A package for performing Peters-Belson with Prognostic Heterogeneity

Performs Peters-Belson with Prognostic Heterogeneity on the data.
}

