% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pblm.R
\name{print.summary.pblm}
\alias{print.summary.pblm}
\title{Print `summary.pblm` objects correctly}
\usage{
\method{print}{summary.pblm}(x, ...)
}
\arguments{
\item{x}{A `summary.pblm` object.}

\item{...}{Additional arguments to `print`.}
}
\value{
Outputs a printed result similar to `print.summary.lm`.
}
\description{
Print `summary.pblm` objects correctly
}
\author{
Josh Errickson
}

