% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{quad}
\alias{quad}
\title{Solve quadratic formula}
\usage{
quad(a, b, c)
}
\arguments{
\item{a}{Coefficient on second-order term.}

\item{b}{Coefficient on first-order term.}

\item{c}{Coefficient on the constant.}
}
\value{
Zeroes of the function a*x^2 + b*x + c = 0.
}
\description{
Given an expression a*x^2 + b*x + c = 0, solve for x.
}
\author{
Josh Errickson
}

