% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{salesdata}
\alias{salesdata}
\title{Sales data with old and new technique}
\format{
A data.frame with 238 rows and 4 columns.
\itemize{
  \item newtechnique Indicator of whether the new technique is used (the treatment variable).
  \item experience Years of experience of salesperson.
  \item previoussales The number of previous sales made in the calendar year.
  \item sale Indicator of whether the sale was successful.
}
}
\usage{
salesdata
}
\description{
This generated data set simulates some data about sales, testing whether a new technique increased sales (a binary response).
}
\keyword{dataset}
