% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.R
\name{sandwich}
\alias{sandwich}
\title{Sandwich estimator allowing clustering}
\usage{
sandwich(x, bread. = bread, meat. = meat, cluster = NULL, ...)
}
\arguments{
\item{x}{a fitted model object.}

\item{bread.}{either a bread matrix or a function for computing
this via \code{bread.(x)}.}

\item{meat.}{either a meat matrix or a function for computing this
via \code{meat.(x)}.}

\item{cluster}{A variable identifying cluster.}

\item{...}{Additional arguments to \code{meat}.}
}
\value{
A covariance \code{matrix}.
}
\description{
This \code{sandwich} function is an overloaded function of
\code{sandwich} from the \code{sandwich} package which enables a
\code{clusters} argument.
}
\seealso{
sandwich::sandwich
}
