% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pblm.R
\name{summary.pblm}
\alias{summary.pblm}
\title{Summary for pblm object}
\usage{
\method{summary}{pblm}(object, ...)
}
\arguments{
\item{object}{An object of class `pblm`.}

\item{...}{Additional arguments to `summary.lm`.}
}
\value{
A summary
}
\description{
Similar to `summary.lm`.
}
\author{
Josh Errickson
s
}

