% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pblm-internal.R
\name{testinverse}
\alias{testinverse}
\title{(Internal) Computes a confidence interval for eta via test
inversion.}
\usage{
testinverse(object, level = 0.95)
}
\arguments{
\item{object}{A pblm object.}

\item{level}{Confidence level. Default is 95\%.}
}
\value{
Vector of length two with the lower and upper bounds.
}
\description{
(Internal) Computes a confidence interval for eta via test
inversion.
}
\author{
Josh Errickson
}

