% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pblm.R
\name{vcov.pblm}
\alias{vcov.pblm}
\title{Returns covariance matrix calculated via sandwich estimation.}
\usage{
\method{vcov}{pblm}(object)
}
\arguments{
\item{object}{pblm object.}
}
\value{
A covariance matrix.
}
\description{
Returns a covariance matrix. Computed using enhanced PB
methods. The variance for pred should NOT be used directly in
confidence intervals.
}
\author{
Josh Errickson
}

