% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbph.R
\name{vcov.pbph}
\alias{vcov.pbph}
\title{Returns covariance matrix calculated via sandwich estimation.}
\usage{
\method{vcov}{pbph}(object)
}
\arguments{
\item{object}{\code{pbph} object.}
}
\value{
A covariance \code{matrix}.
}
\description{
Returns a covariance \code{matrix}. Computed using PBPH
methods. The variance for pred should NOT be used directly in
confidence intervals.
}
