% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{addIntercept}
\alias{addIntercept}
\title{Internal function to add a column for the intercept to a matrix.}
\usage{
addIntercept(x)
}
\arguments{
\item{x}{A \code{matrix} or \code{data.frame}.}
}
\value{
A \code{matrix} (if \code{x} is a \code{matrix}) or
  \code{data.frame} (if \code{x} is a \code{data.frame})
  guaranteed to have an intercept column.
}
\description{
Given a \code{matrix} or \code{data.frame} of size nxp, returns an
nx(p+1) object of the same type where the first column is an
intercept if the intercept is not already in the object (defined
as a first column being all 1's).
}
