% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bread_and_meat.R
\name{bread_and_meat}
\alias{bread11}
\alias{bread21}
\alias{bread22}
\alias{bread_and_meat}
\alias{meat11}
\alias{meat22}
\title{(Internal) Piece-wise generation of Bread and Meat matrices.}
\usage{
bread11(model)

bread22(model)

meat11(model, cluster = NULL)

meat22(model, cluster = NULL)

bread21(eta, resp, covs, pred, treatment)
}
\arguments{
\item{model}{For Bread & Meat that can be calculated using sandwich package,
we only need the model, either first or second stage, depending.}

\item{cluster}{For Meat only; A vector defining clusters.}

\item{eta}{Estimated version of the coefficient on the interaction between
predicted and treatment. This could be from a model or a hypothesis.}

\item{resp}{Vector of responses.}

\item{covs}{Data frame of covariates.}

\item{pred}{Predicted values from first stage.}

\item{treatment}{Vector of 0/1 treatment indicators.}
}
\description{
Computes the Bread and Meat matricies. The diagonal elements are typical
sandwich estimators, with scaling, and thus only use the \code{model}
argument. The off-diagonal Bread element requires further specification.
}

