% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pblm.R
\name{confint.pblm}
\alias{confint.pblm}
\title{Confidence intervals for pblm object}
\usage{
\method{confint}{pblm}(object, parm, level = 0.95, ..., wald.style = FALSE,
  forceDisplayConfInt = FALSE)
}
\arguments{
\item{object}{An object of class \code{pblm}.}

\item{parm}{Parameters}

\item{level}{Confidence level.}

\item{...}{Additional arguments to \code{confint.lm}.}

\item{wald.style}{Logical. Defaults to \code{FALSE} so the method
performs test inversion to obtain proper coverage. If
\code{TRUE}, generates wald-style CI's, which will likely suffer
from undercoverage.}

\item{forceDisplayConfInt}{Logical. Defaults to \code{FALSE}. If a
hypothesis test for the interaction term fails, a confidence
interval is unreliable (and ultimately unnecessary) and thus not
returned. Set this to \code{TRUE} to force computation of the
interval regardless of the hypothesis test.}
}
\value{
Confidence intervals
}
\description{
Similar to \code{confint.lm}.
}

