% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bread_and_meat.R
\name{createBreadAndMeat}
\alias{createBreadAndMeat}
\title{(Internal) Computes Bread and Meat matrices.}
\usage{
createBreadAndMeat(object, clusters = list())
}
\arguments{
\item{object}{A pblm object.}

\item{clusters}{A list of cluster variables to pass to meat.}
}
\value{
A list of b11, b22, m11, and m22.
}
\description{
Computes the pieces of the Bread and Meat which do not depend on eta. (e.g.
all but B21)
}

