% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bread_and_meat.R
\name{glmScale}
\alias{glmScale}
\title{(Internal) Return scaling factor for GLM models}
\usage{
glmScale(model, newdata)
}
\arguments{
\item{model}{A model (lm or glm).}

\item{newdata}{Data to generate the scale.}
}
\value{
Scaling factor. 1 if lm.
}
\description{
(Internal) Return scaling factor for GLM models
}

