% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{makeSaveMatrix}
\alias{makeSaveMatrix}
\title{Generate an empty matrix suitable for saving simulation results.}
\usage{
makeSaveMatrix(colNames, reps)
}
\arguments{
\item{colNames}{A \code{vector} of column names.}

\item{reps}{Total reps.}
}
\value{
An empty \code{matrix}.
}
\description{
Given a \code{vector} of column names and a number of \code{reps}, generates an
appropriately sized \code{matrix} with given column names.
}

