% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.R
\name{meat}
\alias{meat}
\title{Clustered meat matrix for a sandwich estimator}
\usage{
meat(x, adjust = FALSE, clusters = list(), ...)
}
\arguments{
\item{x}{a fitted model object.}

\item{adjust}{See `sandwich::meat`.}

\item{clusters}{A list of variables to cluster on.}

\item{...}{Additional arguments to `sandwich::estfun`.}
}
\value{
A meat matrix
}
\description{
This `meat` function is an overloaded function of `meat` from the
`sandwich` package which enables a `clusters` argument.
}

