% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.R
\name{meat}
\alias{meat}
\title{Clustered meat matrix for a sandwich estimator}
\usage{
meat(x, adjust = FALSE, cluster = NULL, ...)
}
\arguments{
\item{x}{a fitted model object.}

\item{adjust}{See \code{sandwich::meat}. If a cluster is given,
this argument is ignored as the adjustment is always used.}

\item{cluster}{A vector identifying cluster membership}

\item{...}{Additional arguments to \code{sandwich::estfun}.}
}
\value{
A meat \code{matrix}.
}
\description{
This \code{meat} function is an overloaded function of \code{meat}
from the \code{sandwich} package which enables a \code{clusters}
argument.
}

