% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.R
\name{meat}
\alias{meat}
\title{Clustered meat matrix for a sandwich estimator}
\usage{
meat(x, adjust = FALSE, cluster = NULL, ...)
}
\arguments{
\item{x}{a fitted model object.}

\item{adjust}{See \code{sandwich::meat}. If a cluster is given,
this argument is ignored as the adjustment is always used.
The adjustment is \deqn{m/(m - 1) * (n - 1)/(n - k),} where
\eqn{n} is the number of observations, \eqn{k} the number of
estimated parameters, and \eqn{m} the number of clusters. Note
that if there are no clusters (which can be thought of as
each observation being its own cluster), \eqn{m = n} and this
reduces to \deqn{n/(n-k),} the default in \code{sandwich::meat}.
See Cameron, A Colin, and Douglas L Miller. 2010. "Robust
Inference with Clustered Data." Working Papers, University
adjustment used when clustering.
of California, Department of Economics.}

\item{cluster}{A vector identifying cluster membership}

\item{...}{Additional arguments to \code{sandwich::estfun}.}
}
\value{
A meat \code{matrix}.
}
\description{
This \code{meat} function is an overloaded function of \code{meat}
from the \code{sandwich} package which enables a \code{clusters}
argument.
}
\seealso{
sandwich::meat
}
