% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pblm.R
\name{pblm}
\alias{pblm}
\title{Enhanced Peters-Belson method}
\usage{
pblm(mod1, treatment, data, center = TRUE, clusters = list())
}
\arguments{
\item{mod1}{First stage fitted model.}

\item{treatment}{Vector of 0/1 treatment indicators.}

\item{data}{Data where variables in `form` live.}

\item{center}{Default TRUE. Should the predicted values be centered in the
second stage?}

\item{clusters}{List of clusters}
}
\value{
A pblm object which extends lm. Can be passed to `summary` or
  `confint`.

  The return contains an additional object, `epb`, which `lm` doesn't
  include. `epb` is a list which contains several pieces necessary for
  `summary` and `confint` support.
}
\description{
Performs enhanced Peters-Belson on the data.
}

