% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.R
\name{sandwich}
\alias{sandwich}
\title{Sandwich estimator allowing clustering}
\usage{
sandwich(x, bread. = bread, meat. = meat, clusters = list(), ...)
}
\arguments{
\item{x}{a fitted model object.}

\item{bread.}{either a bread matrix or a function for computing
this via `bread.(x)`.}

\item{meat.}{either a meat matrix or a function for computing
this via `meat.(x)`.}

\item{clusters}{A list of variables to cluster on.}

\item{...}{Additional arguments to `meat`.}
}
\value{
A covariance matrix.
}
\description{
This `sandwich` function is an overloaded function of `sandwich`
from the `sandwich` package which enables a `clusters` argument.
}

